Update Broken Pipes.sh

+++ ports/brokenpipes/Broken Pipes.sh
@@ -30,7 +30,7 @@ cd $GAMEDIR
 > "$GAMEDIR/log.txt" && exec > >(tee "$GAMEDIR/log.txt") 2>&1
 
 # Rename data.win
-mv $GAMEDIR/data.win $GAMEDIR/game.droid
+[ -f "./gamedata/data.win" ] && mv $GAMEDIR/data.win $GAMEDIR/game.droid
 
 $ESUDO chmod 666 /dev/uinput
